Clarify docs for repo option and simplify default
The `remote` option should not be used in determining the default for the `repo` option.  If users want to push to the origin remote of the current dir, they can use the default.  If they want to push somewhere else, they use the `repo` option.

The `remote` option is only used if two conditions are met:
 * the `clone` option is set to an existing directory
 * the `clone` directory already has a configured remote (i.e. it wasn't created by this task)