Remove documentation for confusing `remote` option (closes #15)
This option is more misleading than useful.  I tried to update the docs to describe exactly when it should be used, but the use case is so narrow and the documentation so confusion, that I think it is better off left out of the docs entirely.  If someone comes along looking for the feature, it can be added back to the docs.